requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
reportlab==4.0.8
pytest==7.4.3
pytest-cov==4.1.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
structlog==23.2.0
pyOpenSSL==23.3.0
